+++ src/elementosInterfaz_imagen.cpp
@@ -6,7 +6,7 @@ ElementoImagen::ElementoImagen (Gosu::Graphics & g, string ruta, double z, Anima
     imagen.reset(new Gosu::Image(g, Gosu::widen(ruta)));
 }
 
-void ElementoImagen::drawEnd (int x, int y, double z, int a)
+void ElementoImagen::drawEnd (float x, float y, double z, int a)
 {
     //lDEBUG << "img @ " << x << "," << y << "," << z << "(" << a;
     imagen -> draw(x,y,z,1,1, Gosu::Color(a, 255, 255, 255));
